server/requirements.txt
fastapi[all]
uvicorn
passlib
sqlalchemy
pymysql
python-jose
python-dotenv
httpx
pytest
cryptography
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib